requirements.txt
pandas
numpy
scikit-surprise
streamlit
